+++ src/app/features/entities/entity-social/entity-social.component.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 import { Component, OnInit, Input } from '@angular/core';
 
 import { Observable, Subscription, timer } from 'rxjs';
@@ -6,12 +7,13 @@ import Swal from 'sweetalert2';
 
 import { AuthService } from '@auth/auth.service';
 import { IEvent } from '@models/event';
-import { IEventSocial } from '@models/event-social';
+import { IItemSocial } from '@models/item-social';
 // import { CommentType, IComment } from '@models/comment';
 import { IUser } from '@models/user';
 // import { CommentsService } from '@services/comments.service';
-import { EventSocialService } from '@services/events-social.service';
+import { ItemSocialService } from '@services/items-social.service';
 import { UserService } from '@services/users.service';
+import { BaseType } from '@models/base';
 
 // import { CommentsDialogComponent } from '@shared/components/comments-dialog/comments-dialog.component';
 
@@ -26,7 +28,7 @@ export class EntitySocialComponent implements OnInit {
   @Input() userLogged: IUser;
 
   // public comments$: Observable<IComment[]>;
-  public itemSocial: IEventSocial;
+  public itemSocial: IItemSocial;
   public isFav = false;
   // public applause = false;
 
@@ -43,7 +45,7 @@ export class EntitySocialComponent implements OnInit {
     public dialog: MatDialog,
     public authSvc: AuthService,
     // private commentSrv: CommentsService,
-    private socialSrv: EventSocialService,
+    private itemSocialSrv: ItemSocialService,
     private userSrv: UserService,
   ) {
     this.dialogConfig.disableClose = true;
@@ -56,8 +58,8 @@ export class EntitySocialComponent implements OnInit {
 
     // this.comments$ = this.commentSrv.getAllComments(this.entity.id);
 
-    this.socialSrv.getEventSocial(this.entity.id)
-    .subscribe( (itemSocial: IEventSocial) => {
+    this.itemSocialSrv.getItemSocial(this.entity.id)
+    .subscribe( (itemSocial: IItemSocial) => {
         this.itemSocial = itemSocial;
     });
 
@@ -77,38 +79,6 @@ export class EntitySocialComponent implements OnInit {
 
   }
 
-  /*
-  public viewComments(nComments: number): void {
-
-    const userRole = this.userLogged?.role ?? '';
-
-    if ( nComments > 0 || ['SUPER', 'ADMIN', 'AUTOR'].includes(userRole)) {
-
-        this.dialogConfig.width = '600px';
-        this.dialogConfig.height = '600px';
-        this.dialogConfig.data = {
-          itemId: this.entity.id,
-          UserUid: this.userLogged?.uid ?? '',
-          UserName: this.userLogged?.displayName ?? '',
-          UserImage: this.userLogged?.photoURL ?? '',
-          UserRole: userRole,
-          EntityId: ( this.userLogged?.entityDefault?.id ?? '' ),
-          EntityName: ( this.userLogged?.entityDefault?.name ?? '' ),
-          EntityImage: ( this.userLogged?.entityDefault?.imagePath ?? '' ),
-          commentType: CommentType.Event,
-        };
-
-        const dialogRef = this.dialog.open(CommentsDialogComponent, this.dialogConfig);
-      } else {
-        // Swal.fire({
-        //   icon: 'warning',
-        //   title: 'No hay comentarios en este evento',
-        //   text: 'Todos los comentarios son escritos por administradores de la agenda, evitando así spam o malos entendidos',
-        //   confirmButtonColor: '#003A59',
-        // });
-      }
-  }
-  */
 
   public isFavorite(isFav: boolean): void {
 
@@ -118,37 +88,21 @@ export class EntitySocialComponent implements OnInit {
     this.isFav = !this.isFav;
     if ( isFav ) {
       this.userLogged.favEntities.push(this.entity.id);
-      this.socialSrv.addFavourite(this.itemSocial, this.entity.name, this.userLogged.uid, this.userLogged.displayName);
+      this.itemSocialSrv.addFavourite(this.itemSocial.id, BaseType.ENTITY, this.entity.name, this.userLogged.uid, this.userLogged.displayName);
       Swal.fire({
         icon: 'success',
-        title: 'Este evento se ha convertido en uno de tus favoritos',
+        title: 'Esta entidad se ha convertido en uno de tus favoritas',
         confirmButtonColor: '#003A59',
       });
     } else {
-      this.socialSrv.removeFavourite(this.itemSocial, this.entity.name, this.userLogged.uid, this.userLogged.displayName);
+      this.itemSocialSrv.removeFavourite(this.itemSocial.id, BaseType.ENTITY, this.entity.name, this.userLogged.uid, this.userLogged.displayName);
       Swal.fire({
         icon: 'success',
-        title: 'Este evento ha dejado de estar entre tus favoritos',
+        title: 'Esta entidad ha dejado de estar entre tus favoritas',
         confirmButtonColor: '#003A59',
       });
     }
 
     this.userSrv.updateUser(this.userLogged);
   }
-
-  /*
-  public clap(applause: boolean): void {
-    if ( !applause ) {
-      console.log(`applause!`);
-      this.applause = true;
-      this.socialSrv.addClaps(this.itemSocial);
-      const source = timer(3000);
-      const subsTimer$ = source.subscribe(val => {
-        console.log(val);
-        this.applause = false;
-      });
-      this.listOfObservers.push( subsTimer$ );
-    }
-  }
-  */
 }