+++ src/app/features/events/event-social/event-social.component.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 import { Component, OnInit, Input } from '@angular/core';
 
 import { Observable, Subscription, timer } from 'rxjs';
@@ -6,14 +7,15 @@ import Swal from 'sweetalert2';
 
 import { AuthService } from '@auth/auth.service';
 import { IEvent } from '@models/event';
-import { IEventSocial } from '@models/event-social';
+import { IItemSocial } from '@models/item-social';
 import { CommentType, IComment } from '@models/comment';
 import { IUser } from '@models/user';
 import { CommentsService } from '@services/comments.service';
-import { EventSocialService } from '@services/events-social.service';
+import { ItemSocialService } from '@services/items-social.service';
 import { UserService } from '@services/users.service';
 
 import { CommentsDialogComponent } from '@shared/components/comments-dialog/comments-dialog.component';
+import { BaseType } from '@models/base';
 
 @Component({
   selector: 'app-event-social',
@@ -26,7 +28,7 @@ export class EventSocialComponent implements OnInit {
   @Input() userLogged: IUser;
 
   public comments$: Observable<IComment[]>;
-  public eventSocial: IEventSocial;
+  public itemSocial: IItemSocial;
   public isFav = false;
   public applause = false;
 
@@ -43,7 +45,7 @@ export class EventSocialComponent implements OnInit {
     public dialog: MatDialog,
     public authSvc: AuthService,
     private commentSrv: CommentsService,
-    private eventSocialSrv: EventSocialService,
+    private itemSocialSrv: ItemSocialService,
     private userSrv: UserService,
   ) {
     this.dialogConfig.disableClose = true;
@@ -56,9 +58,9 @@ export class EventSocialComponent implements OnInit {
 
     this.comments$ = this.commentSrv.getAllComments(this.event.id);
 
-    this.eventSocialSrv.getEventSocial(this.event.id)
-    .subscribe( (eventSocial: IEventSocial) => {
-        this.eventSocial = eventSocial;
+    this.itemSocialSrv.getItemSocial(this.event.id)
+    .subscribe( (itemSocial: IItemSocial) => {
+        this.itemSocial = itemSocial;
     });
 
     const subs1$ = this.authSvc.afAuth.user
@@ -109,22 +111,24 @@ export class EventSocialComponent implements OnInit {
       }
   }
 
-  public isFavorite(isFav: boolean): void {
+  public setEntityFav(isFav: boolean): void {
 
     this.userLogged.favEvents = this.userLogged.favEvents ?? [];
     this.userLogged.favEvents = this.userLogged.favEvents.filter( (eventId: string) => eventId !== this.event.id );
 
-    this.isFav = !this.isFav;
+    const itemName = this.event.name;
+    const itemId = this.event.id;
+
     if ( isFav ) {
-      this.userLogged.favEvents.push(this.event.id);
-      this.eventSocialSrv.addFavourite(this.eventSocial, this.event.name, this.userLogged.uid, this.userLogged.displayName);
+      this.userLogged.favEvents.push(itemId);
+      this.itemSocialSrv.addFavourite(itemId, BaseType.EVENT, itemName, this.userLogged.uid, this.userLogged.displayName);
       Swal.fire({
         icon: 'success',
         title: 'Este evento se ha convertido en uno de tus favoritos',
         confirmButtonColor: '#003A59',
       });
     } else {
-      this.eventSocialSrv.removeFavourite(this.eventSocial, this.event.name, this.userLogged.uid, this.userLogged.displayName);
+      this.itemSocialSrv.removeFavourite(itemId, BaseType.EVENT, itemName, this.userLogged.uid, this.userLogged.displayName);
       Swal.fire({
         icon: 'success',
         title: 'Este evento ha dejado de estar entre tus favoritos',
@@ -141,7 +145,7 @@ export class EventSocialComponent implements OnInit {
       const userUid = this.userLogged?.uid ?? '';
       const userDisplayName = this.userLogged?.displayName ?? '';
       this.applause = true;
-      this.eventSocialSrv.addClaps(this.eventSocial, this.event.name,  userUid, userDisplayName);
+      this.itemSocialSrv.addClaps(this.itemSocial, BaseType.EVENT, this.event.name,  userUid, userDisplayName);
       const source = timer(3000);
       const subsTimer$ = source.subscribe(val => {
         this.applause = false;